requirements.txt
onnxruntime
numpy
pillow
tqdm
pickle-mixin
streamlit
